Add non-blocking API
sub blocking { 0 }
sub set_privatestorage { ... }
sub get_privatestorage { ... }

This is done by registering hooks with the DJabberd core hook system, then
calling the hooks.

The store and load functions for blocking mode are kept working as
implementations of set_privatestorage and get_privatestorage that
call the callback inline.

(note: this commit with fixes rolled in)